# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Stem Studio API","description":"本机单用户音乐转录 API；模型由独立 worker 执行。","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/system/status":{"get":{"summary":"System Status","operationId":"system_status_api_v1_system_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemStatus"}}}}}}},"/api/v1/projects":{"get":{"summary":"List Projects","operationId":"list_projects_api_v1_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectList"}}}}}},"post":{"summary":"Create Project","operationId":"create_project_api_v1_projects_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProject"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}":{"get":{"summary":"Get Project","operationId":"get_project_api_v1_projects__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Save Project","operationId":"save_project_api_v1_projects__project_id__put","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveProject"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/audio":{"post":{"summary":"Upload Audio","operationId":"upload_audio_api_v1_projects__project_id__audio_post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_audio_api_v1_projects__project_id__audio_post"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/{job_id}":{"get":{"summary":"Get Job","operationId":"get_job_api_v1_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Job"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/{job_id}/cancel":{"post":{"summary":"Cancel Job","operationId":"cancel_job_api_v1_jobs__job_id__cancel_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Job"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/{job_id}/events":{"get":{"summary":"Events","operationId":"events_api_v1_jobs__job_id__events_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}},{"name":"last-event-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last-Event-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/audio/original":{"get":{"summary":"Original Audio","operationId":"original_audio_api_v1_projects__project_id__audio_original_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/tracks/{track_id}/audio":{"get":{"summary":"Track Audio","operationId":"track_audio_api_v1_projects__project_id__tracks__track_id__audio_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}},{"name":"track_id","in":"path","required":true,"schema":{"type":"string","title":"Track Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/export.mid":{"get":{"summary":"Project Midi","operationId":"project_midi_api_v1_projects__project_id__export_mid_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/tracks/{track_id}/export.mid":{"get":{"summary":"Track Midi","operationId":"track_midi_api_v1_projects__project_id__tracks__track_id__export_mid_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}},{"name":"track_id","in":"path","required":true,"schema":{"type":"string","title":"Track Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_audio_api_v1_projects__project_id__audio_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"mode":{"type":"string","enum":["demucs_basic_pitch","transcribe_only"],"title":"Mode","default":"demucs_basic_pitch"},"separationModel":{"type":"string","enum":["htdemucs","htdemucs_6s"],"title":"Separationmodel","default":"htdemucs"},"device":{"type":"string","enum":["auto","cpu","cuda"],"title":"Device","default":"auto"},"transcriptionModel":{"type":"string","enum":["basic_pitch","piano_highres"],"title":"Transcriptionmodel","default":"basic_pitch"},"drumTranscription":{"type":"string","enum":["none","cymbal_onsets"],"title":"Drumtranscription","default":"none"},"bpm":{"anyOf":[{"type":"number","maximum":300.0,"minimum":20.0},{"type":"null"}],"title":"Bpm"}},"type":"object","required":["file"],"title":"Body_upload_audio_api_v1_projects__project_id__audio_post"},"CreateProject":{"properties":{"title":{"type":"string","maxLength":200,"minLength":1,"title":"Title"},"bpm":{"type":"number","maximum":300.0,"minimum":20.0,"title":"Bpm","default":120},"timeSignature":{"type":"string","const":"4/4","title":"Timesignature","default":"4/4"}},"additionalProperties":false,"type":"object","required":["title"],"title":"CreateProject"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Job":{"properties":{"id":{"type":"string","title":"Id"},"projectId":{"type":"string","title":"Projectid"},"status":{"type":"string","enum":["queued","running","succeeded","failed","cancelled"],"title":"Status"},"stage":{"type":"string","enum":["queued","decode","separate","transcribe","merge","done","failed","cancelled"],"title":"Stage"},"progress":{"anyOf":[{"type":"number","maximum":1.0,"minimum":0.0},{"type":"null"}],"title":"Progress"},"message":{"type":"string","title":"Message"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"createdAt":{"type":"string","title":"Createdat"},"updatedAt":{"type":"string","title":"Updatedat"}},"additionalProperties":false,"type":"object","required":["id","projectId","status","stage","progress","message","createdAt","updatedAt"],"title":"Job"},"Note":{"properties":{"id":{"type":"string","maxLength":100,"minLength":1,"title":"Id"},"pitch":{"type":"integer","maximum":127.0,"minimum":0.0,"title":"Pitch"},"startSeconds":{"type":"number","maximum":86400.0,"minimum":0.0,"title":"Startseconds"},"durationSeconds":{"type":"number","maximum":86400.0,"exclusiveMinimum":0.0,"title":"Durationseconds"},"velocity":{"type":"integer","maximum":127.0,"minimum":1.0,"title":"Velocity"}},"additionalProperties":false,"type":"object","required":["id","pitch","startSeconds","durationSeconds","velocity"],"title":"Note"},"Project":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"status":{"type":"string","enum":["empty","queued","processing","ready","failed","cancelled"],"title":"Status"},"bpm":{"type":"number","title":"Bpm"},"timeSignature":{"type":"string","const":"4/4","title":"Timesignature"},"durationSeconds":{"type":"number","title":"Durationseconds"},"revision":{"type":"integer","title":"Revision"},"createdAt":{"type":"string","title":"Createdat"},"updatedAt":{"type":"string","title":"Updatedat"},"sourceAudioUrl":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sourceaudiourl"},"latestJobId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Latestjobid"},"tracks":{"items":{"$ref":"#/components/schemas/Track"},"type":"array","title":"Tracks"},"warnings":{"items":{"type":"string"},"type":"array","title":"Warnings"}},"additionalProperties":false,"type":"object","required":["id","title","status","bpm","timeSignature","durationSeconds","revision","createdAt","updatedAt"],"title":"Project"},"ProjectList":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Project"},"type":"array","title":"Items"}},"additionalProperties":false,"type":"object","required":["items"],"title":"ProjectList"},"SaveProject":{"properties":{"expectedRevision":{"type":"integer","minimum":0.0,"title":"Expectedrevision"},"title":{"anyOf":[{"type":"string","maxLength":200,"minLength":1},{"type":"null"}],"title":"Title"},"bpm":{"anyOf":[{"type":"number","maximum":300.0,"minimum":20.0},{"type":"null"}],"title":"Bpm"},"timeSignature":{"anyOf":[{"type":"string","const":"4/4"},{"type":"null"}],"title":"Timesignature"},"tracks":{"items":{"$ref":"#/components/schemas/Track"},"type":"array","maxItems":64,"title":"Tracks"}},"additionalProperties":false,"type":"object","required":["expectedRevision","tracks"],"title":"SaveProject"},"ServiceCapabilities":{"properties":{"transcribeOnly":{"type":"boolean","title":"Transcribeonly","default":false},"separate4":{"type":"boolean","title":"Separate4","default":false},"separate6":{"type":"boolean","title":"Separate6","default":false},"pianoTranscription":{"type":"boolean","title":"Pianotranscription","default":false},"cymbalOnsets":{"type":"boolean","title":"Cymbalonsets","default":false}},"additionalProperties":false,"type":"object","title":"ServiceCapabilities"},"SystemStatus":{"properties":{"status":{"type":"string","enum":["ready","starting","unavailable"],"title":"Status"},"apiReady":{"type":"boolean","title":"Apiready","default":true},"workerReady":{"type":"boolean","title":"Workerready","default":false},"capabilities":{"$ref":"#/components/schemas/ServiceCapabilities"},"message":{"type":"string","title":"Message"}},"additionalProperties":false,"type":"object","required":["status","message"],"title":"SystemStatus"},"Track":{"properties":{"id":{"type":"string","pattern":"^[a-zA-Z0-9_-]{1,100}$","title":"Id"},"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"kind":{"type":"string","enum":["vocal","drums","bass","other","piano","guitar","strings"],"title":"Kind"},"program":{"type":"integer","maximum":127.0,"minimum":0.0,"title":"Program"},"isDrum":{"type":"boolean","title":"Isdrum","default":false},"muted":{"type":"boolean","title":"Muted","default":false},"solo":{"type":"boolean","title":"Solo","default":false},"volume":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Volume","default":1},"pan":{"type":"number","maximum":1.0,"minimum":-1.0,"title":"Pan","default":0},"audioUrl":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Audiourl"},"transcriptionStatus":{"type":"string","enum":["completed","unsupported","failed"],"title":"Transcriptionstatus","default":"completed"},"notes":{"items":{"$ref":"#/components/schemas/Note"},"type":"array","maxItems":100000,"title":"Notes"},"warnings":{"items":{"type":"string"},"type":"array","title":"Warnings"},"analysis":{"anyOf":[{"$ref":"#/components/schemas/TrackAnalysis"},{"type":"null"}]},"transcriptionEngine":{"anyOf":[{"type":"string","enum":["basic_pitch","piano_highres","cymbal_onsets"]},{"type":"null"}],"title":"Transcriptionengine"}},"additionalProperties":false,"type":"object","required":["id","name","kind","program"],"title":"Track"},"TrackAnalysis":{"properties":{"rmsDbfs":{"type":"number","title":"Rmsdbfs"},"peakDbfs":{"type":"number","title":"Peakdbfs"},"relativeRmsDb":{"type":"number","title":"Relativermsdb"},"lowSignal":{"type":"boolean","title":"Lowsignal"},"autoTranscriptionSkipped":{"type":"boolean","title":"Autotranscriptionskipped"}},"additionalProperties":false,"type":"object","required":["rmsDbfs","peakDbfs","relativeRmsDb","lowSignal","autoTranscriptionSkipped"],"title":"TrackAnalysis","description":"Measured source audio levels; this is not instrument confidence."},"UploadAccepted":{"properties":{"project":{"$ref":"#/components/schemas/Project"},"job":{"$ref":"#/components/schemas/Job"}},"additionalProperties":false,"type":"object","required":["project","job"],"title":"UploadAccepted"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}